requirements.txt
aiogram
python-dotenv
wikipedia